Tweak TaskViewModal, AsigneeSelector, and LabelSelector.

+++ src/containers/Workplace/TaskViewModal/LabelSelector/LabelSelector.tsx
@@ -37,7 +37,8 @@ const useStyles = makeStyles((theme: Theme) =>
     createStyles({
         popover: {
             '& .MuiPopover-paper': {
-                borderRadius: 0
+                borderRadius: 0,
+                marginTop: 5
             }
         },
         root: {
@@ -51,7 +52,12 @@ const useStyles = makeStyles((theme: Theme) =>
             justifyContent: 'flex-start',
             padding: "5px 0 5px 10px",
             borderRadius: 0,
-
+            border: "2px solid white",
+            
+            '&.open': {
+                borderRadius: 3,
+                border: "2px solid " + theme.palette.primary.main,
+            },
             '& p': {
                 fontSize: 14,
                 fontWeight: 300,
@@ -147,6 +153,10 @@ const LabelSelector: React.FC<ILabelSeletorProps> = props => {
     }
 
     const handleDeleteLabel = (label: Label) => {
+        const c = window.confirm("Delete label?");
+
+        if (!c) return;
+
         props.deleteLabel(label.id);
         handlelistItemClick(label, true);
 
@@ -180,7 +190,7 @@ const LabelSelector: React.FC<ILabelSeletorProps> = props => {
 
             <div ref={elemRef} className={classes.root}>
                 <Button
-                    className={classes.button}
+                    className={[classes.button, Boolean(anchorEl) ? "open" : ""].join(' ')}
                     onClick={(event: React.MouseEvent<HTMLButtonElement>) => {
                         setAnchorEl(event.currentTarget);
                     }}
@@ -189,7 +199,7 @@ const LabelSelector: React.FC<ILabelSeletorProps> = props => {
                         {
                             props.task.labels && props.task.labels.length > 0
                             ? _.orderBy(props.task.labels).join(",")
-                            : "No label"
+                            : "None"
                         }
                     </Typography>
                 </Button>
@@ -212,8 +222,8 @@ const LabelSelector: React.FC<ILabelSeletorProps> = props => {
             >
                 <List
                     style={{
-                        width: !props.fullScreen ? (listWidth || 'auto') : 'auto',
-                        maxWidth: props.fullScreen ? (listWidth || 'auto') : 'auto'
+                        width: listWidth,//!props.fullScreen ? (listWidth || 'auto') : 'auto',
+                        maxWidth: listWidth,//props.fullScreen ? (listWidth || 'auto') : 'auto'
                     }}
                 >
                     {
